+++ packages/widgetbook_generator/lib/models/widgetbook_text_scale_factor_data.freezed.dart
@@ -1,0 +1,157 @@
+// coverage:ignore-file
+// GENERATED CODE - DO NOT MODIFY BY HAND
+// ignore_for_file: type=lint
+// ignore_for_file: unused_element, deprecated_member_use, deprecated_member_use_from_same_package, use_function_type_syntax_for_parameters, unnecessary_const, avoid_init_to_null, invalid_override_different_default_values_named, prefer_expression_function_bodies, annotate_overrides, invalid_annotation_target
+
+part of 'widgetbook_text_scale_factor_data.dart';
+
+// **************************************************************************
+// FreezedGenerator
+// **************************************************************************
+
+T _$identity<T>(T value) => value;
+
+final _privateConstructorUsedError = UnsupportedError(
+    'It seems like you constructed your class using `MyClass._()`. This constructor is only meant to be used by freezed and you are not supposed to need it nor use it.\nPlease check the documentation here for more information: https://github.com/rrousselGit/freezed#custom-getters-and-methods');
+
+WidgetbookTextScaleFactorData _$WidgetbookTextScaleFactorDataFromJson(
+    Map<String, dynamic> json) {
+  return _WidgetbookTextScaleFactorData.fromJson(json);
+}
+
+/// @nodoc
+mixin _$WidgetbookTextScaleFactorData {
+  double get value => throw _privateConstructorUsedError;
+
+  Map<String, dynamic> toJson() => throw _privateConstructorUsedError;
+  @JsonKey(ignore: true)
+  $WidgetbookTextScaleFactorDataCopyWith<WidgetbookTextScaleFactorData>
+      get copyWith => throw _privateConstructorUsedError;
+}
+
+/// @nodoc
+abstract class $WidgetbookTextScaleFactorDataCopyWith<$Res> {
+  factory $WidgetbookTextScaleFactorDataCopyWith(
+          WidgetbookTextScaleFactorData value,
+          $Res Function(WidgetbookTextScaleFactorData) then) =
+      _$WidgetbookTextScaleFactorDataCopyWithImpl<$Res>;
+  $Res call({double value});
+}
+
+/// @nodoc
+class _$WidgetbookTextScaleFactorDataCopyWithImpl<$Res>
+    implements $WidgetbookTextScaleFactorDataCopyWith<$Res> {
+  _$WidgetbookTextScaleFactorDataCopyWithImpl(this._value, this._then);
+
+  final WidgetbookTextScaleFactorData _value;
+  // ignore: unused_field
+  final $Res Function(WidgetbookTextScaleFactorData) _then;
+
+  @override
+  $Res call({
+    Object? value = freezed,
+  }) {
+    return _then(_value.copyWith(
+      value: value == freezed
+          ? _value.value
+          : value // ignore: cast_nullable_to_non_nullable
+              as double,
+    ));
+  }
+}
+
+/// @nodoc
+abstract class _$WidgetbookTextScaleFactorDataCopyWith<$Res>
+    implements $WidgetbookTextScaleFactorDataCopyWith<$Res> {
+  factory _$WidgetbookTextScaleFactorDataCopyWith(
+          _WidgetbookTextScaleFactorData value,
+          $Res Function(_WidgetbookTextScaleFactorData) then) =
+      __$WidgetbookTextScaleFactorDataCopyWithImpl<$Res>;
+  @override
+  $Res call({double value});
+}
+
+/// @nodoc
+class __$WidgetbookTextScaleFactorDataCopyWithImpl<$Res>
+    extends _$WidgetbookTextScaleFactorDataCopyWithImpl<$Res>
+    implements _$WidgetbookTextScaleFactorDataCopyWith<$Res> {
+  __$WidgetbookTextScaleFactorDataCopyWithImpl(
+      _WidgetbookTextScaleFactorData _value,
+      $Res Function(_WidgetbookTextScaleFactorData) _then)
+      : super(_value, (v) => _then(v as _WidgetbookTextScaleFactorData));
+
+  @override
+  _WidgetbookTextScaleFactorData get _value =>
+      super._value as _WidgetbookTextScaleFactorData;
+
+  @override
+  $Res call({
+    Object? value = freezed,
+  }) {
+    return _then(_WidgetbookTextScaleFactorData(
+      value: value == freezed
+          ? _value.value
+          : value // ignore: cast_nullable_to_non_nullable
+              as double,
+    ));
+  }
+}
+
+/// @nodoc
+@JsonSerializable()
+class _$_WidgetbookTextScaleFactorData
+    implements _WidgetbookTextScaleFactorData {
+  _$_WidgetbookTextScaleFactorData({required this.value});
+
+  factory _$_WidgetbookTextScaleFactorData.fromJson(
+          Map<String, dynamic> json) =>
+      _$$_WidgetbookTextScaleFactorDataFromJson(json);
+
+  @override
+  final double value;
+
+  @override
+  String toString() {
+    return 'WidgetbookTextScaleFactorData(value: $value)';
+  }
+
+  @override
+  bool operator ==(dynamic other) {
+    return identical(this, other) ||
+        (other.runtimeType == runtimeType &&
+            other is _WidgetbookTextScaleFactorData &&
+            const DeepCollectionEquality().equals(other.value, value));
+  }
+
+  @JsonKey(ignore: true)
+  @override
+  int get hashCode =>
+      Object.hash(runtimeType, const DeepCollectionEquality().hash(value));
+
+  @JsonKey(ignore: true)
+  @override
+  _$WidgetbookTextScaleFactorDataCopyWith<_WidgetbookTextScaleFactorData>
+      get copyWith => __$WidgetbookTextScaleFactorDataCopyWithImpl<
+          _WidgetbookTextScaleFactorData>(this, _$identity);
+
+  @override
+  Map<String, dynamic> toJson() {
+    return _$$_WidgetbookTextScaleFactorDataToJson(this);
+  }
+}
+
+abstract class _WidgetbookTextScaleFactorData
+    implements WidgetbookTextScaleFactorData {
+  factory _WidgetbookTextScaleFactorData({required final double value}) =
+      _$_WidgetbookTextScaleFactorData;
+
+  factory _WidgetbookTextScaleFactorData.fromJson(Map<String, dynamic> json) =
+      _$_WidgetbookTextScaleFactorData.fromJson;
+
+  @override
+  double get value => throw _privateConstructorUsedError;
+  @override
+  @JsonKey(ignore: true)
+  _$WidgetbookTextScaleFactorDataCopyWith<_WidgetbookTextScaleFactorData>
+      get copyWith => throw _privateConstructorUsedError;
+}