+++ packages/widgetbook_generator/lib/models/widgetbook_text_scale_factor_data.g.dart
@@ -1,0 +1,19 @@
+// GENERATED CODE - DO NOT MODIFY BY HAND
+
+part of 'widgetbook_text_scale_factor_data.dart';
+
+// **************************************************************************
+// JsonSerializableGenerator
+// **************************************************************************
+
+_$_WidgetbookTextScaleFactorData _$$_WidgetbookTextScaleFactorDataFromJson(
+        Map<String, dynamic> json) =>
+    _$_WidgetbookTextScaleFactorData(
+      value: (json['value'] as num).toDouble(),
+    );
+
+Map<String, dynamic> _$$_WidgetbookTextScaleFactorDataToJson(
+        _$_WidgetbookTextScaleFactorData instance) =>
+    <String, dynamic>{
+      'value': instance.value,
+    };