+++ packages/widgetbook_generator/lib/models/widgetbook_use_case_data.dart
@@ -1,0 +1,31 @@
+import 'package:freezed_annotation/freezed_annotation.dart';
+import 'package:widgetbook_generator/models/widgetbook_data.dart';
+
+part 'widgetbook_use_case_data.freezed.dart';
+part 'widgetbook_use_case_data.g.dart';
+
+@freezed
+class WidgetbookUseCaseData extends WidgetbookData
+    with _$WidgetbookUseCaseData {
+  factory WidgetbookUseCaseData({
+    // Name of the builder function defining the use-case
+    required String name,
+    // Name of the use-case, e.g. 'Default'
+    required String useCaseName,
+    // Name of the use-case, e.g. 'ElevatedButton'
+    // This will be extracted from the type
+    required String componentName,
+    // Import statement of the use-case definition
+    required String importStatement,
+    // Import statement of the component
+    required String componentImportStatement,
+    required List<String> dependencies,
+    // The path to the file containing the component
+    required String componentDefinitionPath,
+    // The path to the file containing the use-case definition
+    required String useCaseDefinitionPath,
+  }) = _WidgetbookUseCaseData;
+
+  factory WidgetbookUseCaseData.fromJson(Map<String, dynamic> json) =>
+      _$WidgetbookUseCaseDataFromJson(json);
+}