+++ packages/widgetbook_generator/lib/models/widgetbook_use_case_data.freezed.dart
@@ -1,0 +1,314 @@
+// coverage:ignore-file
+// GENERATED CODE - DO NOT MODIFY BY HAND
+// ignore_for_file: type=lint
+// ignore_for_file: unused_element, deprecated_member_use, deprecated_member_use_from_same_package, use_function_type_syntax_for_parameters, unnecessary_const, avoid_init_to_null, invalid_override_different_default_values_named, prefer_expression_function_bodies, annotate_overrides, invalid_annotation_target
+
+part of 'widgetbook_use_case_data.dart';
+
+// **************************************************************************
+// FreezedGenerator
+// **************************************************************************
+
+T _$identity<T>(T value) => value;
+
+final _privateConstructorUsedError = UnsupportedError(
+    'It seems like you constructed your class using `MyClass._()`. This constructor is only meant to be used by freezed and you are not supposed to need it nor use it.\nPlease check the documentation here for more information: https://github.com/rrousselGit/freezed#custom-getters-and-methods');
+
+WidgetbookUseCaseData _$WidgetbookUseCaseDataFromJson(
+    Map<String, dynamic> json) {
+  return _WidgetbookUseCaseData.fromJson(json);
+}
+
+/// @nodoc
+mixin _$WidgetbookUseCaseData {
+  String get name => throw _privateConstructorUsedError;
+  String get useCaseName => throw _privateConstructorUsedError;
+  String get componentName => throw _privateConstructorUsedError;
+  String get importStatement => throw _privateConstructorUsedError;
+  String get componentImportStatement => throw _privateConstructorUsedError;
+  List<String> get dependencies => throw _privateConstructorUsedError;
+  String get componentDefinitionPath => throw _privateConstructorUsedError;
+  String get useCaseDefinitionPath => throw _privateConstructorUsedError;
+
+  Map<String, dynamic> toJson() => throw _privateConstructorUsedError;
+  @JsonKey(ignore: true)
+  $WidgetbookUseCaseDataCopyWith<WidgetbookUseCaseData> get copyWith =>
+      throw _privateConstructorUsedError;
+}
+
+/// @nodoc
+abstract class $WidgetbookUseCaseDataCopyWith<$Res> {
+  factory $WidgetbookUseCaseDataCopyWith(WidgetbookUseCaseData value,
+          $Res Function(WidgetbookUseCaseData) then) =
+      _$WidgetbookUseCaseDataCopyWithImpl<$Res>;
+  $Res call(
+      {String name,
+      String useCaseName,
+      String componentName,
+      String importStatement,
+      String componentImportStatement,
+      List<String> dependencies,
+      String componentDefinitionPath,
+      String useCaseDefinitionPath});
+}
+
+/// @nodoc
+class _$WidgetbookUseCaseDataCopyWithImpl<$Res>
+    implements $WidgetbookUseCaseDataCopyWith<$Res> {
+  _$WidgetbookUseCaseDataCopyWithImpl(this._value, this._then);
+
+  final WidgetbookUseCaseData _value;
+  // ignore: unused_field
+  final $Res Function(WidgetbookUseCaseData) _then;
+
+  @override
+  $Res call({
+    Object? name = freezed,
+    Object? useCaseName = freezed,
+    Object? componentName = freezed,
+    Object? importStatement = freezed,
+    Object? componentImportStatement = freezed,
+    Object? dependencies = freezed,
+    Object? componentDefinitionPath = freezed,
+    Object? useCaseDefinitionPath = freezed,
+  }) {
+    return _then(_value.copyWith(
+      name: name == freezed
+          ? _value.name
+          : name // ignore: cast_nullable_to_non_nullable
+              as String,
+      useCaseName: useCaseName == freezed
+          ? _value.useCaseName
+          : useCaseName // ignore: cast_nullable_to_non_nullable
+              as String,
+      componentName: componentName == freezed
+          ? _value.componentName
+          : componentName // ignore: cast_nullable_to_non_nullable
+              as String,
+      importStatement: importStatement == freezed
+          ? _value.importStatement
+          : importStatement // ignore: cast_nullable_to_non_nullable
+              as String,
+      componentImportStatement: componentImportStatement == freezed
+          ? _value.componentImportStatement
+          : componentImportStatement // ignore: cast_nullable_to_non_nullable
+              as String,
+      dependencies: dependencies == freezed
+          ? _value.dependencies
+          : dependencies // ignore: cast_nullable_to_non_nullable
+              as List<String>,
+      componentDefinitionPath: componentDefinitionPath == freezed
+          ? _value.componentDefinitionPath
+          : componentDefinitionPath // ignore: cast_nullable_to_non_nullable
+              as String,
+      useCaseDefinitionPath: useCaseDefinitionPath == freezed
+          ? _value.useCaseDefinitionPath
+          : useCaseDefinitionPath // ignore: cast_nullable_to_non_nullable
+              as String,
+    ));
+  }
+}
+
+/// @nodoc
+abstract class _$WidgetbookUseCaseDataCopyWith<$Res>
+    implements $WidgetbookUseCaseDataCopyWith<$Res> {
+  factory _$WidgetbookUseCaseDataCopyWith(_WidgetbookUseCaseData value,
+          $Res Function(_WidgetbookUseCaseData) then) =
+      __$WidgetbookUseCaseDataCopyWithImpl<$Res>;
+  @override
+  $Res call(
+      {String name,
+      String useCaseName,
+      String componentName,
+      String importStatement,
+      String componentImportStatement,
+      List<String> dependencies,
+      String componentDefinitionPath,
+      String useCaseDefinitionPath});
+}
+
+/// @nodoc
+class __$WidgetbookUseCaseDataCopyWithImpl<$Res>
+    extends _$WidgetbookUseCaseDataCopyWithImpl<$Res>
+    implements _$WidgetbookUseCaseDataCopyWith<$Res> {
+  __$WidgetbookUseCaseDataCopyWithImpl(_WidgetbookUseCaseData _value,
+      $Res Function(_WidgetbookUseCaseData) _then)
+      : super(_value, (v) => _then(v as _WidgetbookUseCaseData));
+
+  @override
+  _WidgetbookUseCaseData get _value => super._value as _WidgetbookUseCaseData;
+
+  @override
+  $Res call({
+    Object? name = freezed,
+    Object? useCaseName = freezed,
+    Object? componentName = freezed,
+    Object? importStatement = freezed,
+    Object? componentImportStatement = freezed,
+    Object? dependencies = freezed,
+    Object? componentDefinitionPath = freezed,
+    Object? useCaseDefinitionPath = freezed,
+  }) {
+    return _then(_WidgetbookUseCaseData(
+      name: name == freezed
+          ? _value.name
+          : name // ignore: cast_nullable_to_non_nullable
+              as String,
+      useCaseName: useCaseName == freezed
+          ? _value.useCaseName
+          : useCaseName // ignore: cast_nullable_to_non_nullable
+              as String,
+      componentName: componentName == freezed
+          ? _value.componentName
+          : componentName // ignore: cast_nullable_to_non_nullable
+              as String,
+      importStatement: importStatement == freezed
+          ? _value.importStatement
+          : importStatement // ignore: cast_nullable_to_non_nullable
+              as String,
+      componentImportStatement: componentImportStatement == freezed
+          ? _value.componentImportStatement
+          : componentImportStatement // ignore: cast_nullable_to_non_nullable
+              as String,
+      dependencies: dependencies == freezed
+          ? _value.dependencies
+          : dependencies // ignore: cast_nullable_to_non_nullable
+              as List<String>,
+      componentDefinitionPath: componentDefinitionPath == freezed
+          ? _value.componentDefinitionPath
+          : componentDefinitionPath // ignore: cast_nullable_to_non_nullable
+              as String,
+      useCaseDefinitionPath: useCaseDefinitionPath == freezed
+          ? _value.useCaseDefinitionPath
+          : useCaseDefinitionPath // ignore: cast_nullable_to_non_nullable
+              as String,
+    ));
+  }
+}
+
+/// @nodoc
+@JsonSerializable()
+class _$_WidgetbookUseCaseData implements _WidgetbookUseCaseData {
+  _$_WidgetbookUseCaseData(
+      {required this.name,
+      required this.useCaseName,
+      required this.componentName,
+      required this.importStatement,
+      required this.componentImportStatement,
+      required final List<String> dependencies,
+      required this.componentDefinitionPath,
+      required this.useCaseDefinitionPath})
+      : _dependencies = dependencies;
+
+  factory _$_WidgetbookUseCaseData.fromJson(Map<String, dynamic> json) =>
+      _$$_WidgetbookUseCaseDataFromJson(json);
+
+  @override
+  final String name;
+  @override
+  final String useCaseName;
+  @override
+  final String componentName;
+  @override
+  final String importStatement;
+  @override
+  final String componentImportStatement;
+  final List<String> _dependencies;
+  @override
+  List<String> get dependencies {
+    // ignore: implicit_dynamic_type
+    return EqualUnmodifiableListView(_dependencies);
+  }
+
+  @override
+  final String componentDefinitionPath;
+  @override
+  final String useCaseDefinitionPath;
+
+  @override
+  String toString() {
+    return 'WidgetbookUseCaseData(name: $name, useCaseName: $useCaseName, componentName: $componentName, importStatement: $importStatement, componentImportStatement: $componentImportStatement, dependencies: $dependencies, componentDefinitionPath: $componentDefinitionPath, useCaseDefinitionPath: $useCaseDefinitionPath)';
+  }
+
+  @override
+  bool operator ==(dynamic other) {
+    return identical(this, other) ||
+        (other.runtimeType == runtimeType &&
+            other is _WidgetbookUseCaseData &&
+            const DeepCollectionEquality().equals(other.name, name) &&
+            const DeepCollectionEquality()
+                .equals(other.useCaseName, useCaseName) &&
+            const DeepCollectionEquality()
+                .equals(other.componentName, componentName) &&
+            const DeepCollectionEquality()
+                .equals(other.importStatement, importStatement) &&
+            const DeepCollectionEquality().equals(
+                other.componentImportStatement, componentImportStatement) &&
+            const DeepCollectionEquality()
+                .equals(other.dependencies, dependencies) &&
+            const DeepCollectionEquality().equals(
+                other.componentDefinitionPath, componentDefinitionPath) &&
+            const DeepCollectionEquality()
+                .equals(other.useCaseDefinitionPath, useCaseDefinitionPath));
+  }
+
+  @JsonKey(ignore: true)
+  @override
+  int get hashCode => Object.hash(
+      runtimeType,
+      const DeepCollectionEquality().hash(name),
+      const DeepCollectionEquality().hash(useCaseName),
+      const DeepCollectionEquality().hash(componentName),
+      const DeepCollectionEquality().hash(importStatement),
+      const DeepCollectionEquality().hash(componentImportStatement),
+      const DeepCollectionEquality().hash(dependencies),
+      const DeepCollectionEquality().hash(componentDefinitionPath),
+      const DeepCollectionEquality().hash(useCaseDefinitionPath));
+
+  @JsonKey(ignore: true)
+  @override
+  _$WidgetbookUseCaseDataCopyWith<_WidgetbookUseCaseData> get copyWith =>
+      __$WidgetbookUseCaseDataCopyWithImpl<_WidgetbookUseCaseData>(
+          this, _$identity);
+
+  @override
+  Map<String, dynamic> toJson() {
+    return _$$_WidgetbookUseCaseDataToJson(this);
+  }
+}
+
+abstract class _WidgetbookUseCaseData implements WidgetbookUseCaseData {
+  factory _WidgetbookUseCaseData(
+      {required final String name,
+      required final String useCaseName,
+      required final String componentName,
+      required final String importStatement,
+      required final String componentImportStatement,
+      required final List<String> dependencies,
+      required final String componentDefinitionPath,
+      required final String useCaseDefinitionPath}) = _$_WidgetbookUseCaseData;
+
+  factory _WidgetbookUseCaseData.fromJson(Map<String, dynamic> json) =
+      _$_WidgetbookUseCaseData.fromJson;
+
+  @override
+  String get name => throw _privateConstructorUsedError;
+  @override
+  String get useCaseName => throw _privateConstructorUsedError;
+  @override
+  String get componentName => throw _privateConstructorUsedError;
+  @override
+  String get importStatement => throw _privateConstructorUsedError;
+  @override
+  String get componentImportStatement => throw _privateConstructorUsedError;
+  @override
+  List<String> get dependencies => throw _privateConstructorUsedError;
+  @override
+  String get componentDefinitionPath => throw _privateConstructorUsedError;
+  @override
+  String get useCaseDefinitionPath => throw _privateConstructorUsedError;
+  @override
+  @JsonKey(ignore: true)
+  _$WidgetbookUseCaseDataCopyWith<_WidgetbookUseCaseData> get copyWith =>
+      throw _privateConstructorUsedError;
+}